Update contentpage.vue
trigger-top	number	undefined	The distance px to top of target to trigger top affix. (if not set, use top prop)
top	                number	undefined	The css top property after trigger top affix. (if not set, use trigger-top prop)
What could have happened here is that you specified only one and the other was set to the same value, which caused the issue. I modified the top value to reflect what i did on my console but i'm not sure if this will really work as intended as i have no proof of testing